+++ service/register.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/gofrs/uuid/v5"
+	"github.com/sirupsen/logrus"
 	etcdv2 "go.etcd.io/etcd/client/v2"
 
 	"github.com/Scalingo/go-utils/errors/v3"
@@ -30,8 +31,6 @@ const (
 // registration every 5 seconds, and the second one will check if the service
 // credentials don't change and notify otherwise.
 func Register(ctx context.Context, service string, host Host) *Registration {
-	log := logger.Get(ctx)
-
 	if !host.Public && len(host.PrivateHostname) == 0 {
 		host.PrivateHostname = host.Hostname
 	}
@@ -44,6 +43,11 @@ func Register(ctx context.Context, service string, host Host) *Registration {
 	if len(host.PrivateHostname) != 0 && len(host.PrivatePorts) == 0 {
 		host.PrivatePorts = host.Ports
 	}
+
+	ctx, log := logger.WithFieldsToCtx(ctx, logrus.Fields{
+		"hostname":     host.Hostname,
+		"service_name": host.Name,
+	})
 
 	uuidV4, _ := uuid.NewV4()
 	hostUUID := fmt.Sprintf("%s-%s", uuidV4.String(), host.PrivateHostname)
@@ -86,12 +90,14 @@ func Register(ctx context.Context, service string, host Host) *Registration {
 			registration.signalFailure(err)
 			return
 		}
+		log.Info("Service registered in etcd")
 
 		err = ensureInitialHostRegistration(ctx, service, hostKey, hostValue, false)
 		if err != nil {
 			registration.signalFailure(err)
 			return
 		}
+		log.Info("Host registered in etcd")
 
 		publicCredentialsChan <- Credentials{
 			User:     serviceInfos.User,